requirements-dev.txt
# 🏆 Intelligent Retail Analytics Engine v3.0 - Development Requirements
# Comprehensive development and testing dependencies

# ============================================================================
# TESTING FRAMEWORKS
# ============================================================================

pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-xdist==3.5.0
pytest-benchmark==4.0.0
pytest-html==4.1.1
pytest-metadata==3.0.0

# ============================================================================
# CODE QUALITY & LINTING
# ============================================================================

black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
ruff==0.1.6

# ============================================================================
# SECURITY SCANNING
# ============================================================================

bandit==1.7.6
safety==2.3.5
trivy==0.48.0

# ============================================================================
# DEVELOPMENT TOOLS
# ============================================================================

jupyter==1.0.0
jupyterlab==4.0.9
ipykernel==6.26.0
notebook==7.0.6

# ============================================================================
# DEBUGGING & PROFILING
# ============================================================================

debugpy==1.8.0
line-profiler==4.1.2
memory-profiler==0.61.0
py-spy==0.3.14

# ============================================================================
# API TESTING & DOCUMENTATION
# ============================================================================

httpx==0.25.2
requests-toolbelt==1.0.0
openapi-generator-cli==6.6.0

# ============================================================================
# DATABASE DEVELOPMENT
# ============================================================================

pgcli==4.0.1
mycli==1.27.0
litecli==1.9.0

# ============================================================================
# MONITORING & LOGGING (DEVELOPMENT)
# ============================================================================

rich==13.7.0
tqdm==4.66.1
colorama==0.4.6

# ============================================================================
# DEPLOYMENT & CONTAINERIZATION
# ============================================================================

docker==6.1.3
docker-compose==1.29.2
kubernetes==28.1.0

# ============================================================================
# CLOUD DEVELOPMENT
# ============================================================================

google-cloud-sdk==447.0.0
awscli==1.29.57
azure-cli==2.53.1

# ============================================================================
# PERFORMANCE TESTING
# ============================================================================

locust==2.17.0
k6==0.47.0

# ============================================================================
# DOCUMENTATION
# ============================================================================

mkdocs==1.5.3
mkdocs-material==9.4.8
mkdocstrings==0.23.0
mkdocstrings-python==1.7.5

# ============================================================================
# VERSION CONTROL & CI/CD
# ============================================================================

gitpython==3.1.40
PyGithub==2.1.1

# ============================================================================
# ENVIRONMENT MANAGEMENT
# ============================================================================

python-dotenv==1.0.0
pip-tools==7.3.0

# ============================================================================
# MOCKING & FAKING
# ============================================================================

faker==20.1.0
responses==0.24.1
freezegun==1.2.2

# ============================================================================
# CODE COVERAGE & REPORTING
# ============================================================================

coverage==7.3.2
coveralls==3.3.1
codecov==2.1.13

# ============================================================================
# TYPE HINTS & VALIDATION
# ============================================================================

types-all==1.0.0
types-requests==2.31.0
types-redis==4.6.0
types-pytz==2023.3.0

# ============================================================================
# DEVELOPMENT DATABASES
# ============================================================================

sqlite3
psycopg2-binary==2.9.9

# ============================================================================
# IDE SUPPORT
# ============================================================================

jedi==0.19.1
parso==0.8.3
jedi-language-server==0.41.1